Move parser-like test into parser.
